+++ phpstan.neon.dist
@@ -30,6 +30,7 @@ parameters:
         # Deprecated classes
         - 'src/Sylius/Bundle/CoreBundle/Application/Kernel.php'
         - 'src/Sylius/Bundle/UserBundle/Security/UserPasswordEncoder.php'
+        - 'src/Sylius/Bundle/CoreBundle/Console/Command/ShowAvailablePluginsCommand.php'
 
         # To investigate, occurs after release of doctrine/orm 2.14, the processing of these classes ends with exit code 143
         - 'src/Sylius/Bundle/CoreBundle/Doctrine/DQL/**.php'
