+++ src/Sylius/Bundle/AdminBundle/Console/Command/CreateAdminUserCommand.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Sylius\Bundle\AdminBundle\Console\Command;
 
 use Sylius\Bundle\AdminBundle\Console\Command\Factory\QuestionFactoryInterface;
-use Sylius\Bundle\AdminBundle\Command\CreateAdminUser;
+use Sylius\Bundle\AdminBundle\Message\CreateAdminUser;
 use Sylius\Bundle\AdminBundle\Exception\CreateAdminUserFailedException;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
