+++ src/Sylius/Bundle/CoreBundle/Command/InstallCommand.php
@@ -13,124 +13,18 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\Command;
 
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\Process\Exception\RuntimeException;
+trigger_deprecation(
+    'sylius/core-bundle',
+    '1.13',
+    'The "%s" class is deprecated and will be removed in Sylius 2.0. Use "%s" instead.',
+    InstallCommand::class,
+    \Sylius\Bundle\CoreBundle\Console\Command\InstallCommand::class,
+);
 
-final class InstallCommand extends AbstractInstallCommand
-{
-    protected static $defaultName = 'sylius:install';
+class_exists(\Sylius\Bundle\CoreBundle\Console\Command\InstallCommand::class);
 
-    private array $commands = [
-        [
-            'command' => 'check-requirements',
-            'message' => 'Checking system requirements.',
-        ],
-        [
-            'command' => 'database',
-            'message' => 'Setting up the database.',
-        ],
-        [
-            'command' => 'setup',
-            'message' => 'Shop configuration.',
-        ],
-        [
-            'command' => 'jwt-setup',
-            'message' => 'Configuring JWT token.',
-        ],
-        [
-            'command' => 'assets',
-            'message' => 'Installing assets.',
-        ],
-    ];
-
-    protected function configure(): void
+if (false) {
+    final class InstallCommand
     {
-        $this
-            ->setDescription('Installs Sylius in your preferred environment.')
-            ->setHelp(
-                <<<EOT
-The <info>%command.name%</info> command installs Sylius.
-EOT
-            )
-            ->addOption('fixture-suite', 's', InputOption::VALUE_OPTIONAL, 'Load specified fixture suite during install', null)
-        ;
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output): int
-    {
-        $suite = $input->getOption('fixture-suite');
-
-        $outputStyle = new SymfonyStyle($input, $output);
-        $outputStyle->writeln('<info>Installing Sylius...</info>');
-        $outputStyle->writeln($this->getSyliusLogo());
-
-        $this->ensureDirectoryExistsAndIsWritable((string) $this->getContainer()->getParameter('kernel.cache_dir'), $output);
-
-        $errored = false;
-        /**
-         * @var int $step
-         * @var array $command
-         */
-        foreach ($this->commands as $step => $command) {
-            try {
-                $outputStyle->newLine();
-                $outputStyle->section(sprintf(
-                    'Step %d of %d. <info>%s</info>',
-                    $step + 1,
-                    count($this->commands),
-                    $command['message'],
-                ));
-
-                $parameters = [];
-                if ('database' === $command['command'] && null !== $suite) {
-                    $parameters['--fixture-suite'] = $suite;
-                }
-
-                $this->commandExecutor->runCommand('sylius:install:' . $command['command'], $parameters, $output);
-            } catch (RuntimeException) {
-                $errored = true;
-            }
-        }
-
-        $outputStyle->newLine(2);
-        $outputStyle->success($this->getProperFinalMessage($errored));
-        $outputStyle->writeln('You can now open your store at the following path under the website root: /');
-
-        return $errored ? 1 : 0;
-    }
-
-    private function getProperFinalMessage(bool $errored): string
-    {
-        if ($errored) {
-            return 'Sylius has been installed, but some error occurred.';
-        }
-
-        return 'Sylius has been successfully installed.';
-    }
-
-    private function getSyliusLogo(): string
-    {
-        return '
-           <info>,</info>
-         <info>,;:,</info>
-       <info>`;;;.:`</info>
-      <info>`::;`  :`</info>
-       <info>:::`   `</info>          .\'++:           \'\'.   \'.
-       <info>`:::</info>             :+\',;+\'          :+;  `+.
-        <info>::::</info>            +\'   :\'          `+;
-        <info>`:::,</info>           \'+`     ++    :+.`+; `++. ;+\'    \'\'  ,++++.
-         <info>,:::`</info>          `++\'.   .+:  `+\' `+;  .+,  ;+    +\'  +;  \'\'
-          <info>::::`</info>           ,+++.  \'+` :+. `+;  `+,  ;+    +\'  \'+.
-   <info>,.     .::::</info>             .++` `+: +\'  `+;  `+,  ;+    +\'  `;++;
-<info>`;;.:::`   :::::</info>             :+.  \'+,+.  `+;  `+,  ;+   `+\'     .++
- <info>.;;;;;;::`.::::,</info>       +\'` `++   `++\'   `+;  `+:  :+. `++\'  \'.  ;+
-  <info>,;;;;;;;;;:::::</info>       .+++++`    ;+,    ++;  ++, `\'+++,\'+\' :++++,
-   <info>,;;;;;;;;;:::</info>`                  ;\'
-    <info>:;;;;;;;;;:,</info>                :.:+,
-     <info>;;;;;;;;;:</info>                 ;++,'
-        ;
     }
 }
