+++ src/Sylius/Bundle/CoreBundle/Command/InstallDatabaseCommand.php
@@ -13,53 +13,18 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\Command;
 
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Style\SymfonyStyle;
+trigger_deprecation(
+    'sylius/core-bundle',
+    '1.13',
+    'The "%s" class is deprecated and will be removed in Sylius 2.0. Use "%s" instead.',
+    InstallDatabaseCommand::class,
+    \Sylius\Bundle\CoreBundle\Console\Command\InstallDatabaseCommand::class,
+);
 
-final class InstallDatabaseCommand extends AbstractInstallCommand
-{
-    protected static $defaultName = 'sylius:install:database';
+class_exists(\Sylius\Bundle\CoreBundle\Console\Command\InstallDatabaseCommand::class);
 
-    protected function configure(): void
+if (false) {
+    final class InstallDatabaseCommand
     {
-        $this
-            ->setDescription('Install Sylius database.')
-            ->setHelp(
-                <<<EOT
-The <info>%command.name%</info> command creates Sylius database.
-EOT
-            )
-            ->addOption('fixture-suite', 's', InputOption::VALUE_OPTIONAL, 'Load specified fixture suite during install', null)
-        ;
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output): int
-    {
-        $suite = $input->getOption('fixture-suite');
-
-        $outputStyle = new SymfonyStyle($input, $output);
-        $outputStyle->writeln(sprintf(
-            'Creating Sylius database for environment <info>%s</info>.',
-            $this->getEnvironment(),
-        ));
-
-        $commands = $this
-            ->getContainer()
-            ->get('sylius.commands_provider.database_setup')
-            ->getCommands($input, $output, $this->getHelper('question'))
-        ;
-
-        $this->runCommands($commands, $output);
-        $outputStyle->newLine();
-
-        $parameters = [];
-        if (null !== $suite) {
-            $parameters['--fixture-suite'] = $suite;
-        }
-        $this->commandExecutor->runCommand('sylius:install:sample-data', $parameters, $output);
-
-        return 0;
     }
 }
