+++ src/Sylius/Bundle/CoreBundle/Command/InstallSampleDataCommand.php
@@ -13,75 +13,18 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\Command;
 
-use Symfony\Component\Console\Helper\QuestionHelper;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\ConfirmationQuestion;
-use Symfony\Component\Console\Style\SymfonyStyle;
+trigger_deprecation(
+    'sylius/core-bundle',
+    '1.13',
+    'The "%s" class is deprecated and will be removed in Sylius 2.0. Use "%s" instead.',
+    InstallSampleDataCommand::class,
+    \Sylius\Bundle\CoreBundle\Console\Command\InstallSampleDataCommand::class,
+);
 
-final class InstallSampleDataCommand extends AbstractInstallCommand
-{
-    protected static $defaultName = 'sylius:install:sample-data';
+class_exists(\Sylius\Bundle\CoreBundle\Console\Command\InstallSampleDataCommand::class);
 
-    protected function configure(): void
+if (false) {
+    final class InstallSampleDataCommand
     {
-        $this
-            ->setDescription('Install sample data into Sylius.')
-            ->setHelp(
-                <<<EOT
-The <info>%command.name%</info> command loads the sample data for Sylius.
-EOT
-            )
-            ->addOption('fixture-suite', 's', InputOption::VALUE_OPTIONAL, 'Load specified fixture suite during install', null)
-        ;
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output): int
-    {
-        /** @var QuestionHelper $questionHelper */
-        $questionHelper = $this->getHelper('question');
-        $suite = $input->getOption('fixture-suite');
-
-        $outputStyle = new SymfonyStyle($input, $output);
-        $outputStyle->newLine();
-        $outputStyle->writeln(sprintf(
-            'Loading sample data for environment <info>%s</info> from suite <info>%s</info>.',
-            $this->getEnvironment(),
-            $suite ?? 'default',
-        ));
-        $outputStyle->writeln('<error>Warning! This action will erase your database.</error>');
-
-        if (!$questionHelper->ask($input, $output, new ConfirmationQuestion('Continue? (y/N) ', null !== $suite))) {
-            $outputStyle->writeln('Cancelled loading sample data.');
-
-            return 0;
-        }
-
-        try {
-            $publicDir = $this->getContainer()->getParameter('sylius_core.public_dir');
-
-            $this->ensureDirectoryExistsAndIsWritable($publicDir . '/media/', $output);
-            $this->ensureDirectoryExistsAndIsWritable($publicDir . '/media/image/', $output);
-        } catch (\RuntimeException $exception) {
-            $outputStyle->writeln($exception->getMessage());
-
-            return 1;
-        }
-
-        $parameters = ['--no-interaction' => true];
-
-        if (null !== $suite) {
-            $parameters['suite'] = $suite;
-        }
-
-        $commands = [
-            'sylius:fixtures:load' => $parameters,
-        ];
-
-        $this->runCommands($commands, $output);
-        $outputStyle->newLine(2);
-
-        return 0;
     }
 }
